requirements.txt
# Core
numpy>=1.24.0,<2.0
pandas>=2.0.0
scikit-learn>=1.3.0
pyyaml>=6.0

# ML / Forecasting
prophet>=1.1
statsmodels>=0.14.0
xgboost>=2.0.0
lightgbm>=4.0.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
loguru>=0.7.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
ruff>=0.1.0
